--- .babelrc
@@ -1,4 +1,0 @@
-{
-  "presets": ["es2015", "stage-3"],
-  "plugins": ["transform-runtime"]
-}
