--- circle.yml
@@ -1,7 +1,0 @@
-machine:
-  environment:
-    NPM_CONFIG_PROGRESS: 'false'
-test:
-  override:
-    - nvm use 4 && npm test
-    - nvm use 6 && COVERALLS=1 npm test
