+++ package.json
@@ -11,7 +11,8 @@
     "Christophe Bonello <christophe.bonello@zentrick.com>"
   ],
   "engines": {
-    "node": ">=4"
+    "node": ">=6",
+    "yarn": "*"
   },
   "license": "MIT",
   "files": [
@@ -20,10 +21,13 @@
     "index.js"
   ],
   "scripts": {
-    "clean": "gulp clean",
-    "build": "gulp build",
-    "prepublish": "in-publish && npm run build || not-in-publish",
-    "test": "gulp test"
+    "build": "rimraf lib && babel src -d lib",
+    "prepublish": "in-publish && yarn run build || not-in-publish",
+    "lint": "standard '{src,test,scripts}/**/*.js'",
+    "test:integration": "mocha --require babel-polyfill --compilers js:babel-register 'test/lib/setup.js' 'test/integration/**/*.js'",
+    "test:cover": "cross-env NODE_ENV=test nyc yarn run test:integration",
+    "test": "yarn run lint && yarn run test:cover",
+    "ci": "yarn run lint && MOCHA_FILE=$CIRCLE_TEST_REPORTS/junit.xml yarn run test:cover -- --reporter mocha-junit-reporter"
   },
   "repository": "zentrick/iab-vast-parser",
   "bugs": "https://github.com/zentrick/iab-vast-parser/issues",
@@ -34,31 +38,26 @@
     "xmldom": "^0.1.22"
   },
   "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-es2015": "^6.24.0",
-    "babel-preset-stage-3": "^6.22.0",
+    "babel-cli": "^6.26.0",
+    "babel-plugin-istanbul": "^4.1.5",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
+    "babel-register": "^6.26.0",
     "chai": "^4.1.2",
     "chai-as-promised": "^7.1.1",
-    "del": "^3.0.0",
-    "fs-promise": "^2.0.0",
-    "globule": "^1.0.0",
-    "gulp": "^3.9.1",
-    "gulp-babel": "^7.0.1",
-    "gulp-coveralls": "^0.1.4",
-    "gulp-istanbul": "^1.0.0",
-    "gulp-load-plugins": "^1.5.0",
-    "gulp-mocha": "^3.0.1",
-    "gulp-sourcemaps": "^2.4.1",
-    "gulp-standard": "^10.1.2",
-    "gulp-util": "^3.0.8",
+    "cross-env": "^5.1.3",
+    "fs-extra": "^5.0.0",
+    "globule": "^1.2.0",
+    "hard-rejection": "^1.0.0",
     "in-publish": "^2.0.0",
-    "isparta": "^4.0.0",
     "json-stable-stringify": "^1.0.1",
+    "mocha": "^5.0.2",
+    "mocha-junit-reporter": "^1.17.0",
     "normalize-newline": "^3.0.0",
-    "run-sequence": "^2.2.1",
-    "standard": "^11.0.0",
-    "yargs": "^11.0.0"
+    "nyc": "^11.4.1",
+    "replace-ext": "^1.0.0",
+    "rimraf": "^2.6.2",
+    "standard": "^11.0.0"
   },
   "standard": {
     "globals": [
@@ -67,5 +66,32 @@
       "expect",
       "DOMParser"
     ]
+  },
+  "babel": {
+    "presets": [
+      "env"
+    ],
+    "env": {
+      "test": {
+        "plugins": [
+          "istanbul"
+        ]
+      }
+    }
+  },
+  "nyc": {
+    "check-coverage": true,
+    "lines": 95,
+    "statements": 95,
+    "functions": 95,
+    "branches": 80,
+    "include": [
+      "src/**/*.js"
+    ],
+    "require": [
+      "babel-register"
+    ],
+    "sourceMap": false,
+    "instrument": false
   }
 }
