--- src/.babelrc
@@ -1,3 +1,0 @@
-{
-  "presets": ["es2015"]
-}
