+++ src/factory/resource.js
@@ -2,9 +2,8 @@ import createStaticResource from './static-resource'
 import createIFrameResource from './iframe-resource'
 import createHTMLResource from './html-resource'
 
-export default ($parent) => {
-  return $parent.staticResource != null ? createStaticResource($parent.staticResource)
-    : $parent.iFrameResource != null ? createIFrameResource($parent.iFrameResource)
-    : $parent.htmlResource != null ? createHTMLResource($parent.htmlResource)
-    : null
-}
+export default ($parent) =>
+  ($parent.staticResource != null) ? createStaticResource($parent.staticResource)
+    : ($parent.iFrameResource != null) ? createIFrameResource($parent.iFrameResource)
+      : ($parent.htmlResource != null) ? createHTMLResource($parent.htmlResource)
+        : null
