+++ src/util/is-non-empty-string.js
@@ -1,1 +1,1 @@
- export default (str) => (typeof str === 'string' && str.length > 0)
+export default (str) => (typeof str === 'string' && str.length > 0)
